more comments, fixed sending times

+++ commands/maintanance.h
@@ -11,7 +11,7 @@
 class CmdConfirmPeer: public NetworkCommand {
 public:
     CmdConfirmPeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Confirm Peer");
     }
@@ -26,7 +26,7 @@ public:
 class CmdConfirmHost: public NetworkCommand {
 public:
     CmdConfirmHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Confirm Host");
     }
@@ -40,7 +40,7 @@ public:
 class CmdAskHost: public NetworkCommand {
 public:
     CmdAskHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Ask Host");
     }
@@ -55,7 +55,7 @@ public:
 class CmdAskPeer: public NetworkCommand {
 public:
     CmdAskPeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Ask Peer");
     }
@@ -70,7 +70,7 @@ public:
 class CmdPingPeer: public NetworkCommand {
 public:
     CmdPingPeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Ping Peer");
     }
@@ -84,7 +84,7 @@ public:
 class CmdPingHost: public NetworkCommand {
 public:
     CmdPingHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Ping Host");
     }
@@ -98,7 +98,7 @@ public:
 class CmdGoodbyePeer: public NetworkCommand {
 public:
     CmdGoodbyePeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Goodbye Peer");
     }
@@ -111,7 +111,7 @@ public:
 class CmdGoodbyeHost: public NetworkCommand {
 public:
     CmdGoodbyeHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return("Goodbye Host");
     }
@@ -124,7 +124,7 @@ public:
 class CmdSayGoodbye: public NetworkCommand {
 public:
     CmdSayGoodbye(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t , struct sockaddr_storage &, void *) { return true; }
+    virtual bool execute(int64_t , struct sockaddr_storage &, void *) { return true; }
     virtual void execute();
 
     virtual std::string getName() {
