+++ commands/transfer.h
@@ -11,7 +11,7 @@
 class CmdDistributePeer: public NetworkCommand {
 public:
     CmdDistributePeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Distr Peer");
     }
@@ -26,7 +26,7 @@ public:
 class CmdDistributeHost: public NetworkCommand {
 public:
     CmdDistributeHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Distr Host");
     }
@@ -40,7 +40,7 @@ public:
 class CmdReturnPeer: public NetworkCommand {
 public:
     CmdReturnPeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Return peer");
     }
@@ -53,7 +53,7 @@ public:
 class CmdReturnHost: public NetworkCommand {
 public:
     CmdReturnHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Return host");
     }
@@ -68,7 +68,7 @@ public:
 class CmdGatherNeighborsPeer: public NetworkCommand {
 public:
     CmdGatherNeighborsPeer(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Gather neighbors peer");
     }
@@ -82,7 +82,7 @@ public:
 class CmdGatherNeighborsHost: public NetworkCommand {
 public:
     CmdGatherNeighborsHost(VideoState *st, NetworkHandler *hndl): NetworkCommand(st, hndl) {}
-    virtual bool execute(int32_t fd, struct sockaddr_storage &address, void *data);
+    virtual bool execute(int64_t fd, struct sockaddr_storage &address, void *data);
     virtual std::string getName() {
         return ("Gather neighbors host");
     }
